union all aggregates from bounded context PM into one ProjectAggregate
