--- src/main/kotlin/ru/quipy/api/ParticipantAggregate.kt
@@ -1,7 +1,0 @@
-package ru.quipy.api
-
-import ru.quipy.core.annotations.AggregateType
-import ru.quipy.domain.Aggregate
-
-@AggregateType(aggregateEventsTableName = "aggregate-participant")
-class ParticipantAggregate : Aggregate
