+++ src/main/kotlin/ru/quipy/projections/AnnotationBasedProjectEventsSubscriber.kt
@@ -3,8 +3,7 @@ package ru.quipy.projections
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
-import ru.quipy.api.ProjectAggregate
-import ru.quipy.api.ProjectCreatedEvent
+import ru.quipy.api.*
 import ru.quipy.streams.annotation.AggregateSubscriber
 import ru.quipy.streams.annotation.SubscribeEvent
 
@@ -19,5 +18,25 @@ class AnnotationBasedProjectEventsSubscriber {
     @SubscribeEvent
     fun projectCreatedSubscriber(event: ProjectCreatedEvent) {
         logger.info("Project created: {}", event.title)
+    }
+
+    @SubscribeEvent
+    fun participantAddedToProjectSubscriber(event: ParticipantAddedToProjectEvent) {
+        logger.info("Participant {} added to project", event.participantId)
+    }
+
+    @SubscribeEvent
+    fun statusCreatedSubscriber(event: StatusCreatedEvent) {
+        logger.info("Status created: {}", event.statusName)
+    }
+
+    @SubscribeEvent
+    fun taskCreatedSubscriber(event: TaskCreatedEvent) {
+        logger.info("Task created: {}", event.taskName)
+    }
+
+    @SubscribeEvent
+    fun taskAssigneeAddedEventSubscriber(event: TaskAssigneeAddedEvent) {
+        logger.info("Assignee {} added to task {}", event.participantId, event.taskId)
     }
 }
